Remove redundant stories, reorganize field components
