--- libs/ui/src/components/field-helper/field-helper.spec.tsx
@@ -1,10 +1,0 @@
-import { render } from '@testing-library/react';
-
-import FieldHelper from './field-helper';
-
-describe('Text', () => {
-  it('should render successfully', () => {
-    const { baseElement } = render(<FieldHelper />);
-    expect(baseElement).toBeTruthy();
-  });
-});
