--- libs/ui/src/components/field-label/field-label.spec.tsx
@@ -1,10 +1,0 @@
-import { render } from '@testing-library/react';
-
-import FieldLabel from './field-label';
-
-describe('Text', () => {
-  it('should render successfully', () => {
-    const { baseElement } = render(<FieldLabel />);
-    expect(baseElement).toBeTruthy();
-  });
-});
