--- libs/ui/src/components/field-label/field-label.stories.tsx
@@ -1,15 +1,0 @@
-import type { Meta, StoryObj } from '@storybook/react';
-import FieldLabel, { FieldLabelProps } from './field-label';
-
-const meta: Meta<typeof FieldLabel> = {
-  component: FieldLabel,
-  title: 'FieldLabel',
-};
-export default meta;
-
-export const Primary: StoryObj<FieldLabelProps> = {
-  args: {
-    id: 'label',
-    label: 'Label',
-  },
-};
