--- libs/ui/src/components/field-wrapper/field-wrapper.spec.tsx
@@ -1,10 +1,0 @@
-import { render } from '@testing-library/react';
-
-import FieldWrapper from './field-wrapper';
-
-describe('Text', () => {
-  it('should render successfully', () => {
-    const { baseElement } = render(<FieldWrapper />);
-    expect(baseElement).toBeTruthy();
-  });
-});
