+++ libs/ui/src/components/field/wrapper/index.tsx
@@ -1,7 +1,7 @@
 'use client';
 
-import { cx } from '../../utils';
-import styles from './field-wrapper.module.css';
+import { cx } from '../../../utils';
+import styles from './wrapper.module.css';
 
 export interface FieldWrapperProps {
   children: React.ReactNode;
