+++ libs/ui/src/components/form-error/form-error.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import FormError, { FormErrorProps } from './form-error';
+import FormError, { FormErrorProps } from '.';
 
 const meta: Meta<typeof FormError> = {
   component: FormError,
