+++ libs/ui/src/components/icon/icon.spec.tsx
@@ -1,6 +1,6 @@
 import { render } from '@testing-library/react';
 
-import Icon from './icon';
+import Icon from '.';
 
 describe('Icon', () => {
   it('should render successfully', () => {
