+++ libs/ui/src/components/icon/index.tsx
