+++ libs/ui/src/components/link/link.spec.tsx
@@ -1,6 +1,6 @@
 import { render } from '@testing-library/react';
 
-import Link from './link';
+import Link from '.';
 
 describe('Link', () => {
   it('should render successfully', () => {
