+++ libs/ui/src/components/link/link.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import Link, { LinkProps } from './link';
+import Link, { LinkProps } from '.';
 
 const meta: Meta<typeof Link> = {
   component: Link,
