+++ libs/ui/src/components/password-field/password-field.spec.tsx
@@ -1,6 +1,6 @@
 import { render } from '@testing-library/react';
 
-import PasswordField from './password-field';
+import PasswordField from '.';
 
 describe('Text', () => {
   it('should render successfully', () => {
