+++ libs/ui/src/components/spinner/spinner.spec.tsx
@@ -1,6 +1,6 @@
 import { render } from '@testing-library/react';
 
-import Spinner from './spinner';
+import Spinner from '.';
 
 describe('Spinner', () => {
   it('should render successfully', () => {
