+++ libs/ui/src/components/text/text.spec.tsx
@@ -1,6 +1,6 @@
 import { render } from '@testing-library/react';
 
-import Text from './text';
+import Text from '.';
 
 describe('Text', () => {
   it('should render successfully', () => {
