+++ libs/ui/src/index.ts
@@ -1,9 +1,9 @@
-export * from './components/button/button';
-export * from './components/icon/icon';
-export * from './components/link/link';
-export * from './components/text/text';
-export * from './components/text-field/text-field';
-export * from './components/password-field/password-field';
-export * from './components/checkbox-field/checkbox-field';
-export * from './components/form-error/form-error';
-export * from './components/spinner/spinner';
+export * from './components/button';
+export * from './components/icon';
+export * from './components/link';
+export * from './components/text';
+export * from './components/text-field';
+export * from './components/password-field';
+export * from './components/checkbox-field';
+export * from './components/form-error';
+export * from './components/spinner';
